extract FilesAdapter from FilesPageFragment
